House cleaning

+++ server/controllers/boards.js
@@ -1,9 +1,17 @@
 const sql = require("../dbconfig");
 const sqlstring = require("sqlstring");
+const typeCheck = require('type-check').typeCheck;
 
 module.exports = exports = {
   createBoard: (req, res) => {
     let { acronym, title } = req.body;
+
+    let validInputFlag = typeCheck('[String]',[acronym, title]);
+    if(!validInputFlag){
+      res.status(400);
+      return res.send(new Error('Invalid Input'));
+    }
+    
     sql.query(
       sqlstring.format(`insert into boards(acronym,title) values(?,?);`, [
         acronym,
@@ -25,6 +33,13 @@ module.exports = exports = {
 
   deleteBoard: (req, res) => {
     let { acronym } = req.body;
+
+    let validInputFlag = typeCheck('String',acronym);
+    if(!validInputFlag){
+      res.status(400);
+      return res.send(new Error('Invalid Input'))
+    }
+    
     sql.query(
       sqlstring.format(`delete from boards where acronym = ?;`, [acronym]),
       (error, results, fields) => {
@@ -36,6 +51,13 @@ module.exports = exports = {
 
   updateBoardName: (req, res) => {
     let { acronym, new_acronym } = req.body;
+
+    let validInputFlag = typeCheck('[String]',[acronym,new_acronym]);
+    if(!validInputFlag){
+      res.status(400);
+      return res.send(new Error('Invalid Input'));
+    }
+    
     sql.query(
       sqlstring.format(`update boards set acronym = ? where acronym = ?;`, [
         new_acronym,
@@ -50,6 +72,13 @@ module.exports = exports = {
 
   updateBoardTitle: (req, res) => {
     let { acronym, new_title } = req.body;
+
+    let validInputFlag = typeCheck('[String]',acronym, new_title);
+    if(!validInputFlag){
+      res.status(400);
+      return res.send(new Error('Invalid Input'));
+    }
+    
     sql.query(
       sqlstring.format(`update boards set title = ? where acronym = ?;`, [
         new_title,
